Reorganize tests and add some big model test cases
